feat(rk3588): report actual measured pvtpll clocks
grf block has a simple counter the counts the PVTPLL clock pulses. The
counter has an input clock which can be configured using the CRU block,
but in current implementation it is always XIN=24M. The counter
continuously counts the PVTPLL clock ticks for the last CON1 * input
frequency seconds. In current implementation, CON1 is always set to 24
so total measurement time is 24 / 24M = 1us. Total last number of
samples are stored in the first 14bits of status2 register.

This commit simply waits 1us for counter to pupulate and calculates the
pvtpll frequency out of it. For reference, it is also possible to
increase CON1 counter to bigger values (less than it wont overflow the
14bit counter) to get a more accurate reading, but for the sake of
simplicity i just did a simple 1us measurement with existing fixed
values.

With the new measurement now it is possible to measure with 1Mhz
accuracy how the pvtpll freuqency is doing, instead of blindly reporting
back the set frequency which hides critical information about the chip
degredation and environmental affects over the pvtpll clocks

Signed-off-by: Hüseyin BIYIK <boogiepop@gmx.com>